// pet-store/types/paths/pet/{petId}.types.ts
export type getPetById = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: getPetById_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<
      {
        200: {
          headers: never;
          requiredHeaders: never;
          content: {
            "application/json": {
              schema: Pet;
            };
            "application/xml": {
              schema: Pet;
            };
          };
          examples: {};
        };
        400: {
          headers: never;
          requiredHeaders: never;
          content: never;
          examples: {};
        };
        404: {
          headers: never;
          requiredHeaders: never;
          content: never;
          examples: {};
        };
      } & {
        [statusCode in Exclude<HttpStatusCode, 200 | 400 | 404>]: {
          headers: never;
          requiredHeaders: never;
          content: never;
          examples: {};
        };
      }
    >;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Updates a pet resource based on the form data.
 */
export type updatePetWithForm = (
  $: OmitValueWhenNever<{
    query: updatePetWithForm_Query;
    querystring: never;
    path: updatePetWithForm_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<
      {
        200: {
          headers: never;
          requiredHeaders: never;
          content: {
            "application/json": {
              schema: Pet;
            };
            "application/xml": {
              schema: Pet;
            };
          };
          examples: {};
        };
        400: {
          headers: never;
          requiredHeaders: never;
          content: never;
          examples: {};
        };
      } & {
        [statusCode in Exclude<HttpStatusCode, 200 | 400>]: {
          headers: never;
          requiredHeaders: never;
          content: never;
          examples: {};
        };
      }
    >;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Delete a pet.
 */
export type deletePet = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: deletePet_Path;
    headers: deletePet_Headers;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<
      {
        200: {
          headers: never;
          requiredHeaders: never;
          content: never;
          examples: {};
        };
        400: {
          headers: never;
          requiredHeaders: never;
          content: never;
          examples: {};
        };
      } & {
        [statusCode in Exclude<HttpStatusCode, 200 | 400>]: {
          headers: never;
          requiredHeaders: never;
          content: never;
          examples: {};
        };
      }
    >;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type getPetById_Path = {
  /**
   * ID of pet to return
   */
  petId: number;
};

export type updatePetWithForm_Query = {
  /**
   * Name of pet that needs to be updated
   */
  name?: string;
  /**
   * Status of pet that needs to be updated
   */
  status?: string;
};

export type updatePetWithForm_Path = {
  /**
   * ID of pet that needs to be updated
   */
  petId: number;
};

export type deletePet_Path = {
  /**
   * Pet id to delete
   */
  petId: number;
};

export type deletePet_Headers = { api_key?: string };